Reorganizing some of the major routes to be more modular (phase 1 of cleanup)
